Implemented a mesh modifier
By using our new mesh fuser, we can edit a mesh easily. Allow us to remove and add hex to our mesh at a whim. Additionally, our FusedMesh are faster to construct together than using the combine instance.
We can generate a 100 x 100 grid in .3 seconds compared to the .5, and a 500 x 500 in 10 seconds compared the the 14 seconds. On average 30 - 40% increase in speed