Accurate chunk bounds and map dimensions
Instead of manually calculate the map and chunk bounds,
We can use the bound property supplied by the mesh renderers of the chunks

With this info, our bounds are even more precise.
This has now allowed us to disable chunks that are not in view of the camera.
The SetChunkStatus methods are now tested and proven to be working as intended